Implemented recursion abortion.  Fixed nullptr issue.  Added examples/Makefile.
There was a problem when trying to deep display nullptr.  This problem
is now fixed and a configurable string is displayed for nullptr.

In the course of this fix, I found that the deep display of pointers
does not work at all.  The compiler prefers to consider pointers as an
integer address.  I do not see a way to solve this on ostream-level,
but on second thought it's even better to do dereferncing only inside
containers; after all, the user can write cout << *ptr_to_sth.

Although not possible for containers of the standard library, user
containers may introduce endless recursion by containing a pointer to
itself.  Now, ostrut recognizes cycles and breaks them, outputting a
configurable string to inform about it.